+++ display_ad_data/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DisplayOfAdDataConfig(AppConfig):
-    name = 'display_of_ad_data'
+    name = 'display_ad_data'
